Improve structure and documentation of Query Input connector
